chore: remove search parsing
